restruct xio_session
Split xio_session.c to 3 files xio_session_server.c for
server specific code and xio_session_client for client
specific code. Common code is left in xio_session.c
A private header file xio_session_priv.h was added and
it contains the prototypes of the fuctions which are
private to the session "class". (static in one file implementation)

Signed-off-by: Shlomo Pongratz <shlomop@mellanox.com>