make kmem cache name persistent
Make the name of the cache persistent.
According to the documentation the name must be valid until the cache is
destroyed. The name is used in /proc/slabinfo to identify the cache. and
since the kerned doesn't copy the them but only keeps a pointer to it
the module needs to keep the name persistent.

Reported-by: <xuejiaolong@gmail.com>
Signed-off-by: Shlomo Pongratz <shlomop@mellanox.com>