remove wrong sg table assignement in tbl_copy
In tbl_copy the number on destination nents was wrongly set
to the number of source nents. With xnbd a copy was done from
one buffer of 8k to two buffers of 4k each. Since the number of
nents was set to 1 than one scatter element of 4K was missing,
which caused the next iterator to return a NULL pointer.

Issue: 418468

Change-Id: I230d484b824bd21ab1a49f408c829671cd2edd4d
Signed-off-by: Shlomo Pongratz <shlomop@mellanox.com>